log; maybe_auth
